4.6,星期四,好久没有提交了

+++ xuecheng-plus-content/xuecheng-plus-content-model/src/main/java/tang/content/po/CoursePublish.java
@@ -138,12 +138,12 @@ public class CoursePublish implements Serializable {
     /**
      * 现价
      */
-    private Double price;
+    private Float price;
 
     /**
      * 原价
      */
-    private Double originalPrice;
+    private Float originalPrice;
 
     /**
      * 课程有效期天数
